Moved slash command files to main commands folder
