+++ src/commands/IssuesCommand.ts
@@ -1,11 +1,11 @@
 import {ColorResolvable, CommandInteraction, MessageEmbed} from "discord.js";
 import {Discord, Slash, SlashOption} from "discordx";
-import GitHubService from "../../services/GitHubService";
-import GitHubIssue from "../../interfaces/GitHubIssue";
-import DateUtils from "../../utils/DateUtils";
-import StringUtils from "../../utils/StringUtils";
-import getConfigValue from "../../utils/getConfigValue";
-import GenericObject from "../../interfaces/GenericObject";
+import GitHubService from "../services/GitHubService";
+import GitHubIssue from "../interfaces/GitHubIssue";
+import DateUtils from "../utils/DateUtils";
+import StringUtils from "../utils/StringUtils";
+import getConfigValue from "../utils/getConfigValue";
+import GenericObject from "../interfaces/GenericObject";
 
 @Discord()
 class IssuesCommand {
