+++ src/commands/ResourcesCommand.ts
