+++ src/commands/SearchCommand.ts
@@ -1,8 +1,8 @@
 import {ColorResolvable, CommandInteraction, MessageEmbed} from "discord.js";
 import {Discord, Slash, SlashOption} from "discordx";
-import InstantAnswerService from "../../services/InstantAnswerService";
-import getConfigValue from "../../utils/getConfigValue";
-import GenericObject from "../../interfaces/GenericObject";
+import InstantAnswerService from "../services/InstantAnswerService";
+import getConfigValue from "../utils/getConfigValue";
+import GenericObject from "../interfaces/GenericObject";
 
 @Discord()
 class SearchCommand {
