+++ src/config.json
@@ -15,7 +15,7 @@
     "MEMBER_ROLE_COLOR": "#FFFFFE",
     "PRODUCTION_ENV": "production",
     "DEVELOPMENT_ENV": "dev",
-    "slash_commands_directory": "commands/slash",
+    "commands_directory": "commands",
     "handlers_directory": "event/handlers",
     "LOG_CHANNEL_ID": "405068878151024640",
     "MOD_CHANNEL_ID": "495713774205141022",
