+++ test/commands/InspectCommandTest.ts
@@ -3,9 +3,8 @@ import { expect } from "chai";
 import { Collection, EmbedField, GuildMember, GuildMemberRoleManager, Role, Formatters } from "discord.js";
 import { BaseMocks } from "@lambocreeper/mock-discord.js";
 
-import InspectCommand from "../../../src/commands/slash/InspectCommand";
-import DateUtils from "../../../src/utils/DateUtils";
-import DiscordUtils from "../../../src/utils/DiscordUtils";
+import InspectCommand from "../../src/commands/InspectCommand";
+import DiscordUtils from "../../src/utils/DiscordUtils";
 
 const roleCollection = new Collection([["12345", new Role(BaseMocks.getClient(), {
 	"id": "12345",
